# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: folicular
  description: |
    Backend service for Luteal: anonymous identity, offline-first delta sync,
    research-backed canonical data model, clearly labeled cycle estimates.

    This document is the contract's single source of truth. The Go server is
    validated against it (internal/contract tests); the Android client
    generates its DTOs from it. Prose documentation lives in docs/api.md.

    Conventions: dates are ISO-8601 calendar dates (format: date); instants
    are RFC 3339 UTC (format: date-time). Synchronized record IDs are
    client-generated UUIDv7 strings. Errors are RFC 9457 problem+json.
  version: 1.0.0
  contact:
    name: Luteal
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/duo/view:
    get:
      tags: [duo]
      operationId: duoView
      summary: Role-aware partner projection
      description: |
        Partner: only fields with an active grant; ungranted fields are
        null (absence deliberately indistinguishable from no data).
        Tracker: data fields null, support thread always visible. Private
        notes and raw observations never pass through this endpoint.
      responses:
        "200":
          description: Projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DuoView" }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/support-requests:
    post:
      tags: [duo]
      operationId: createSupportRequest
      summary: Ask for support (either member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [link_id, kind]
              properties:
                link_id: { $ref: "#/components/schemas/UUID" }
                kind: { $ref: "#/components/schemas/SupportKind" }
                message: { type: string, maxLength: 1000 }
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupportRequest" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/duo/support-requests/{requestID}/ack:
    patch:
      tags: [duo]
      operationId: ackSupportRequest
      summary: Acknowledge a support request (recipient only)
      parameters:
        - name: requestID
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "204": { description: Acknowledged }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 256-bit device token issued at registration (ltok_...)
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    UUID:
      type: string
      format: uuid
      example: 019832e0-6c14-7000-8000-000000000001
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    DuoRole:
      type: string
      enum: [tracker, partner]
...
    SupportKind:
      type: string
      enum: [general, comfort, practical, space]
    SupportRequest:
      type: object
      required: [id, author_role, kind, message, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        author_role: { $ref: "#/components/schemas/DuoRole" }
        kind: { $ref: "#/components/schemas/SupportKind" }
        message: { type: string }
        created_at: { type: string, format: date-time }
        acknowledged_at: { type: string, format: date-time, nullable: true }
    DuoView:
      type: object
      required: [link_id, role, as_of]
      properties:
        link_id: { $ref: "#/components/schemas/UUID" }
        role: { $ref: "#/components/schemas/DuoRole" }
        as_of: { type: string, format: date-time }
        cycle_day: { type: integer, nullable: true }
        period_estimate:
          type: object
          nullable: true
          required: [window_start, window_end]
          properties:
            window_start: { type: string, format: date }
            window_end: { type: string, format: date }
        mood: { $ref: "#/components/schemas/SharedLevel" }
        energy: { $ref: "#/components/schemas/SharedLevel" }
        support_requests:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/SupportRequest" }
    SharedLevel:
      type: object
      nullable: true
      required: [date, level]
      properties:
        date: { type: string, format: date }
        level: { type: integer, minimum: 1, maximum: 5 }